+++ apps/web/views/application_layout.rb
@@ -2,6 +2,18 @@ module Web
   module Views
     class ApplicationLayout
       include Web::Layout
+
+      def flash_message(flash)
+        type = flash_message_type(flash)
+        return nil unless type
+
+        flash[type]
+      end
+
+      def flash_message_type(flash)
+        return :notice  if flash[:notice]
+        return :error   if flash[:error]
+      end
     end
   end
 end