+++ spec/lib/concierge/suppliers/rentals_united/commands/property_fetcher_spec.rb
@@ -16,13 +16,12 @@ RSpec.describe RentalsUnited::Commands::PropertyFetcher do
 
     result = subject.fetch_property
     expect(result).not_to be_success
-    expect(result.error.code).to eq("56")
-    expect(result.error.data).to eq("Property does not exist.")
+    expected_error_message = "Response indicating the Status with ID `56`, and description `Property does not exist.`"
+    expect(result.error.code).to eq(:unrecognised_response)
+    expect(result.error.data).to eq(expected_error_message)
 
     event = Concierge.context.events.last.to_h
-    expect(event[:message]).to eq(
-      "Response indicating the Status with ID `56`, and description `Property does not exist.`"
-    )
+    expect(event[:message]).to eq(expected_error_message)
     expect(event[:backtrace]).to be_kind_of(Array)
     expect(event[:backtrace].any?).to be true
   end